+++ src/vs/workbench/contrib/extensions/test/browser/extension.test.ts
